Added AnimateDiff Settings + Adjust PE nodes
