--- animatediff/nodes_experimental.py
@@ -1,143 +1,0 @@
-import torch
-
-from .model_injection import MotionModelSettings
-
-
-class AnimateDiffModelSettingsSimple:
-    @classmethod
-    def INPUT_TYPES(s):
-        return {
-            "required": {
-                "motion_pe_stretch": ("INT", {"default": 0, "min": 0, "step": 1}),
-            },
-            "optional": {
-                "mask_motion_scale": ("MASK",),
-                "min_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
-                "max_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
-            }
-        }
-    
-    RETURN_TYPES = ("MOTION_MODEL_SETTINGS",)
-    CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings/experimental"
-    FUNCTION = "get_motion_model_settings"
-
-    def get_motion_model_settings(self, motion_pe_stretch: int,
-                                  mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
-        motion_model_settings = MotionModelSettings(
-            motion_pe_stretch=motion_pe_stretch,
-            mask_attn_scale=mask_motion_scale,
-            mask_attn_scale_min=min_motion_scale,
-            mask_attn_scale_max=max_motion_scale,
-            )
-
-        return (motion_model_settings,)
-
-
-class AnimateDiffModelSettingsAdvanced:
-    @classmethod
-    def INPUT_TYPES(s):
-        return {
-            "required": {
-                "pe_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "other_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "motion_pe_stretch": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "cap_initial_pe_length": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "interpolate_pe_to_length": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "initial_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "final_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
-            },
-            "optional": {
-                "mask_motion_scale": ("MASK",),
-                "min_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
-                "max_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
-            }
-        }
-    
-    RETURN_TYPES = ("MOTION_MODEL_SETTINGS",)
-    CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings/experimental"
-    FUNCTION = "get_motion_model_settings"
-
-    def get_motion_model_settings(self, pe_strength: float, attn_strength: float, other_strength: float,
-                                  motion_pe_stretch: int,
-                                  cap_initial_pe_length: int, interpolate_pe_to_length: int,
-                                  initial_pe_idx_offset: int, final_pe_idx_offset: int,
-                                  mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
-        motion_model_settings = MotionModelSettings(
-            pe_strength=pe_strength,
-            attn_strength=attn_strength,
-            other_strength=other_strength,
-            cap_initial_pe_length=cap_initial_pe_length,
-            interpolate_pe_to_length=interpolate_pe_to_length,
-            initial_pe_idx_offset=initial_pe_idx_offset,
-            final_pe_idx_offset=final_pe_idx_offset,
-            motion_pe_stretch=motion_pe_stretch,
-            mask_attn_scale=mask_motion_scale,
-            mask_attn_scale_min=min_motion_scale,
-            mask_attn_scale_max=max_motion_scale,
-            )
-
-        return (motion_model_settings,)
-
-
-class AnimateDiffModelSettingsAdvancedAttnStrengths:
-    @classmethod
-    def INPUT_TYPES(s):
-        return {
-            "required": {
-                "pe_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_q_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_k_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_v_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_out_weight_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "attn_out_bias_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "other_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
-                "motion_pe_stretch": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "cap_initial_pe_length": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "interpolate_pe_to_length": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "initial_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
-                "final_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
-            },
-            "optional": {
-                "mask_motion_scale": ("MASK",),
-                "min_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
-                "max_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
-            }
-        }
-    
-    RETURN_TYPES = ("MOTION_MODEL_SETTINGS",)
-    CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings/experimental"
-    FUNCTION = "get_motion_model_settings"
-
-    def get_motion_model_settings(self, pe_strength: float, attn_strength: float,
-                                  attn_q_strength: float,
-                                  attn_k_strength: float,
-                                  attn_v_strength: float,
-                                  attn_out_weight_strength: float,
-                                  attn_out_bias_strength: float,
-                                  other_strength: float,
-                                  motion_pe_stretch: int,
-                                  cap_initial_pe_length: int, interpolate_pe_to_length: int,
-                                  initial_pe_idx_offset: int, final_pe_idx_offset: int,
-                                  mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
-        motion_model_settings = MotionModelSettings(
-            pe_strength=pe_strength,
-            attn_strength=attn_strength,
-            attn_q_strength=attn_q_strength,
-            attn_k_strength=attn_k_strength,
-            attn_v_strength=attn_v_strength,
-            attn_out_weight_strength=attn_out_weight_strength,
-            attn_out_bias_strength=attn_out_bias_strength,
-            other_strength=other_strength,
-            cap_initial_pe_length=cap_initial_pe_length,
-            interpolate_pe_to_length=interpolate_pe_to_length,
-            initial_pe_idx_offset=initial_pe_idx_offset,
-            final_pe_idx_offset=final_pe_idx_offset,
-            motion_pe_stretch=motion_pe_stretch,
-            mask_attn_scale=mask_motion_scale,
-            mask_attn_scale_min=min_motion_scale,
-            mask_attn_scale_max=max_motion_scale,
-            )
-
-        return (motion_model_settings,)
