+++ animatediff/nodes_gen1.py
@@ -4,12 +4,13 @@ import torch
 import comfy.sample as comfy_sample
 from comfy.model_patcher import ModelPatcher
 
+from .ad_settings import AdjustPEGroup, AnimateDiffSettings, AdjustPE
 from .context import ContextOptions, ContextOptionsGroup, ContextSchedules
 from .logger import logger
 from .utils_model import BetaSchedules, get_available_motion_loras, get_available_motion_models, get_motion_lora_path
 from .utils_motion import ADKeyframeGroup, get_combined_multival
 from .motion_lora import MotionLoraInfo, MotionLoraList
-from .model_injection import InjectionParams, ModelPatcherAndInjector, MotionModelGroup, MotionModelSettings, load_motion_lora_as_patches, load_motion_module_gen1, load_motion_module_gen2, validate_model_compatibility_gen2
+from .model_injection import InjectionParams, ModelPatcherAndInjector, MotionModelGroup, load_motion_lora_as_patches, load_motion_module_gen1, load_motion_module_gen2, validate_model_compatibility_gen2
 from .sample_settings import SampleSettings, SeedNoiseGeneration
 from .sampling import motion_sample_factory
 
@@ -27,7 +28,7 @@ class AnimateDiffLoaderGen1:
             "optional": {
                 "context_options": ("CONTEXT_OPTIONS",),
                 "motion_lora": ("MOTION_LORA",),
-                "ad_settings": ("MOTION_MODEL_SETTINGS",),
+                "ad_settings": ("AD_SETTINGS",),
                 "ad_keyframes": ("AD_KEYFRAMES",),
                 "sample_settings": ("SAMPLE_SETTINGS",),
                 "scale_multival": ("MULTIVAL",),
@@ -42,7 +43,7 @@ class AnimateDiffLoaderGen1:
     def load_mm_and_inject_params(self,
         model: ModelPatcher,
         model_name: str, beta_schedule: str,# apply_mm_groupnorm_hack: bool,
-        context_options: ContextOptionsGroup=None, motion_lora: MotionLoraList=None, ad_settings: MotionModelSettings=None,
+        context_options: ContextOptionsGroup=None, motion_lora: MotionLoraList=None, ad_settings: AnimateDiffSettings=None,
         sample_settings: SampleSettings=None, scale_multival=None, effect_multival=None, ad_keyframes: ADKeyframeGroup=None,
     ):
         # load motion module and motion settings, if included
@@ -65,7 +66,7 @@ class AnimateDiffLoaderGen1:
 
         # set motion_scale and motion_model_settings
         if not ad_settings:
-            ad_settings = MotionModelSettings()
+            ad_settings = AnimateDiffSettings()
         ad_settings.attn_scale = 1.0
         params.set_motion_model_settings(ad_settings)
 
@@ -105,7 +106,7 @@ class LegacyAnimateDiffLoaderWithContext:
             "optional": {
                 "context_options": ("CONTEXT_OPTIONS",),
                 "motion_lora": ("MOTION_LORA",),
-                "ad_settings": ("MOTION_MODEL_SETTINGS",),
+                "ad_settings": ("AD_SETTINGS",),
                 "sample_settings": ("SAMPLE_SETTINGS",),
                 "motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
                 "apply_v2_models_properly": ("BOOLEAN", {"default": True}),
@@ -121,7 +122,7 @@ class LegacyAnimateDiffLoaderWithContext:
     def load_mm_and_inject_params(self,
         model: ModelPatcher,
         model_name: str, beta_schedule: str,# apply_mm_groupnorm_hack: bool,
-        context_options: ContextOptionsGroup=None, motion_lora: MotionLoraList=None, ad_settings: MotionModelSettings=None, motion_model_settings: MotionModelSettings=None,
+        context_options: ContextOptionsGroup=None, motion_lora: MotionLoraList=None, ad_settings: AnimateDiffSettings=None, motion_model_settings: AnimateDiffSettings=None,
         sample_settings: SampleSettings=None, motion_scale: float=1.0, apply_v2_models_properly: bool=False, ad_keyframes: ADKeyframeGroup=None,
     ):
         if ad_settings is not None:
@@ -138,7 +139,7 @@ class LegacyAnimateDiffLoaderWithContext:
             params.set_context(context_options)
         # set motion_scale and motion_model_settings
         if not motion_model_settings:
-            motion_model_settings = MotionModelSettings()
+            motion_model_settings = AnimateDiffSettings()
         motion_model_settings.attn_scale = motion_scale
         params.set_motion_model_settings(motion_model_settings)
 
@@ -179,12 +180,152 @@ class AnimateDiffModelSettings:
             }
         }
     
-    RETURN_TYPES = ("MOTION_MODEL_SETTINGS",)
+    RETURN_TYPES = ("AD_SETTINGS",)
     CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings"
     FUNCTION = "get_motion_model_settings"
 
     def get_motion_model_settings(self, mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
-        motion_model_settings = MotionModelSettings(
+        motion_model_settings = AnimateDiffSettings(
+            mask_attn_scale=mask_motion_scale,
+            mask_attn_scale_min=min_motion_scale,
+            mask_attn_scale_max=max_motion_scale,
+            )
+
+        return (motion_model_settings,)
+
+
+class AnimateDiffModelSettingsSimple:
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "motion_pe_stretch": ("INT", {"default": 0, "min": 0, "step": 1}),
+            },
+            "optional": {
+                "mask_motion_scale": ("MASK",),
+                "min_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
+                "max_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
+            }
+        }
+    
+    RETURN_TYPES = ("AD_SETTINGS",)
+    CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings/experimental"
+    FUNCTION = "get_motion_model_settings"
+
+    def get_motion_model_settings(self, motion_pe_stretch: int,
+                                  mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
+        adjust_pe = AdjustPEGroup(AdjustPE(motion_pe_stretch=motion_pe_stretch))
+        motion_model_settings = AnimateDiffSettings(
+            adjust_pe=adjust_pe,
+            motion_pe_stretch=motion_pe_stretch,
+            mask_attn_scale=mask_motion_scale,
+            mask_attn_scale_min=min_motion_scale,
+            mask_attn_scale_max=max_motion_scale,
+            )
+
+        return (motion_model_settings,)
+
+
+class AnimateDiffModelSettingsAdvanced:
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "pe_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "other_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "motion_pe_stretch": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "cap_initial_pe_length": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "interpolate_pe_to_length": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "initial_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "final_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
+            },
+            "optional": {
+                "mask_motion_scale": ("MASK",),
+                "min_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
+                "max_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
+            }
+        }
+    
+    RETURN_TYPES = ("AD_SETTINGS",)
+    CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings/experimental"
+    FUNCTION = "get_motion_model_settings"
+
+    def get_motion_model_settings(self, pe_strength: float, attn_strength: float, other_strength: float,
+                                  motion_pe_stretch: int,
+                                  cap_initial_pe_length: int, interpolate_pe_to_length: int,
+                                  initial_pe_idx_offset: int, final_pe_idx_offset: int,
+                                  mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
+        adjust_pe = AdjustPEGroup(AdjustPE(motion_pe_stretch=motion_pe_stretch,
+                             cap_initial_pe_length=cap_initial_pe_length, interpolate_pe_to_length=interpolate_pe_to_length,
+                             initial_pe_idx_offset=initial_pe_idx_offset, final_pe_idx_offset=final_pe_idx_offset))
+        motion_model_settings = AnimateDiffSettings(
+            adjust_pe=adjust_pe,
+            pe_strength=pe_strength,
+            attn_strength=attn_strength,
+            other_strength=other_strength,
+            mask_attn_scale=mask_motion_scale,
+            mask_attn_scale_min=min_motion_scale,
+            mask_attn_scale_max=max_motion_scale,
+            )
+
+        return (motion_model_settings,)
+
+
+class AnimateDiffModelSettingsAdvancedAttnStrengths:
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "pe_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_q_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_k_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_v_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_out_weight_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "attn_out_bias_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "other_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 10.0, "step": 0.0001}),
+                "motion_pe_stretch": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "cap_initial_pe_length": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "interpolate_pe_to_length": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "initial_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
+                "final_pe_idx_offset": ("INT", {"default": 0, "min": 0, "step": 1}),
+            },
+            "optional": {
+                "mask_motion_scale": ("MASK",),
+                "min_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
+                "max_motion_scale": ("FLOAT", {"default": 1.0, "min": 0.0, "step": 0.001}),
+            }
+        }
+    
+    RETURN_TYPES = ("AD_SETTINGS",)
+    CATEGORY = "Animate Diff 🎭🅐🅓/① Gen1 nodes ①/motion settings/experimental"
+    FUNCTION = "get_motion_model_settings"
+
+    def get_motion_model_settings(self, pe_strength: float, attn_strength: float,
+                                  attn_q_strength: float,
+                                  attn_k_strength: float,
+                                  attn_v_strength: float,
+                                  attn_out_weight_strength: float,
+                                  attn_out_bias_strength: float,
+                                  other_strength: float,
+                                  motion_pe_stretch: int,
+                                  cap_initial_pe_length: int, interpolate_pe_to_length: int,
+                                  initial_pe_idx_offset: int, final_pe_idx_offset: int,
+                                  mask_motion_scale: torch.Tensor=None, min_motion_scale: float=1.0, max_motion_scale: float=1.0):
+        adjust_pe = AdjustPEGroup(AdjustPE(motion_pe_stretch=motion_pe_stretch,
+                             cap_initial_pe_length=cap_initial_pe_length, interpolate_pe_to_length=interpolate_pe_to_length,
+                             initial_pe_idx_offset=initial_pe_idx_offset, final_pe_idx_offset=final_pe_idx_offset))
+        motion_model_settings = AnimateDiffSettings(
+            adjust_pe=adjust_pe,
+            pe_strength=pe_strength,
+            attn_strength=attn_strength,
+            attn_q_strength=attn_q_strength,
+            attn_k_strength=attn_k_strength,
+            attn_v_strength=attn_v_strength,
+            attn_out_weight_strength=attn_out_weight_strength,
+            attn_out_bias_strength=attn_out_bias_strength,
+            other_strength=other_strength,
             mask_attn_scale=mask_motion_scale,
             mask_attn_scale_min=min_motion_scale,
             mask_attn_scale_max=max_motion_scale,
